Add authorization

+++ app/controllers/messages_controller.rb
@@ -3,7 +3,7 @@
 class MessagesController < ApplicationController
   def index
     session[:id] ||= SecureRandom.urlsafe_base64
-    @messages = Message.all
+    @messages = Message.all.order(created_at: :desc)
   end
 
   def new
@@ -14,6 +14,7 @@ class MessagesController < ApplicationController
 
   def create
     @message = Message.new(message_params)
+    @message.user = current_user
     if @message.save
       broadcast_message(@message, :create)
       head :created
@@ -24,12 +25,14 @@ class MessagesController < ApplicationController
 
   def edit
     @message = Message.find(params[:id])
+    authorize(@message, :manage?)
 
     respond_to :js
   end
 
   def update
     @message = Message.find(params[:id])
+    authorize(@message, :manage?)
     if @message.update(message_params)
       broadcast_message(@message, :update)
       head :ok
@@ -40,6 +43,7 @@ class MessagesController < ApplicationController
 
   def destroy
     @message = Message.find(params[:id])
+    authorize(@message, :manage?)
     if @message.destroy
       broadcast_message(@message, :destroy)
       head :no_content
@@ -66,7 +70,7 @@ class MessagesController < ApplicationController
   def html(message)
     ApplicationController.render(
       partial: 'messages/message',
-      locals: { message: message }
+      locals: { message: message, current_user: current_user }
     )
   end
 end
